[refactor] board 관련 엔티티 자잘자잘하게 수정
