+++ src/main/java/com/example/demo/domain/board/domain/entity/BoardCategory.java
@@ -1,7 +1,6 @@
-package com.example.demo.domain.category.domain.entity;
+package com.example.demo.domain.board.domain.entity;
 
 
-import com.example.demo.domain.board.domain.entity.Board;
 import com.example.demo.global.base.domain.BaseEntity;
 import jakarta.persistence.*;
 import lombok.Getter;
@@ -23,20 +22,14 @@ public class BoardCategory extends BaseEntity {
     @JoinColumn(name = "board_id", nullable = false)
     private Board board;
 
-    @ManyToOne(cascade = CascadeType.PERSIST)
-    @JoinColumn(name = "category_id",
-            nullable = false)
+    @ManyToOne
+    @JoinColumn(name = "category_id", nullable = false)
     private Category category;
 
-    private String categoryName;
 
-    public BoardCategory(String categoryName) {
-        this.categoryName = categoryName;
-    }
 
     public BoardCategory(Board board, Category category) {
         this.board = board;
         this.category = category;
-        this.categoryName = category.getName();
     }
 }
