+++ src/main/java/com/example/demo/domain/board/domain/entity/Category.java
@@ -1,5 +1,6 @@
-package com.example.demo.domain.category.domain.entity;
+package com.example.demo.domain.board.domain.entity;
 
+import com.example.demo.domain.board.domain.entity.BoardCategory;
 import com.example.demo.global.base.domain.BaseEntity;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotBlank;
@@ -22,10 +23,7 @@ public class Category extends BaseEntity {
     @NotBlank(message = "이름은 빈 값일 수 없습니다.")
     private String name;
 
-    @OneToMany(mappedBy = "category",
-            fetch = FetchType.LAZY,
-            cascade = CascadeType.ALL,
-            orphanRemoval = true) // TODO: 삭제 전이 한번 다시 볼 필요 있음
+    @OneToMany(mappedBy = "category", fetch = FetchType.LAZY,cascade = CascadeType.PERSIST)
     private List<BoardCategory> boardCategories = new ArrayList<>();
 
     public Category(String name) {
